Round 30 items 2 and 3: full-name dates in four languages, and the banner joins the photo
ITEM 3 — "ne 13 9" was the bug, and abbreviations were the cause.

`short` weekday and month degrade worst in exactly the languages that need them
most: Czech `ne`, Russian `вс`. Full names now, everywhere a game date renders
— the WHEN row, the games-list day headings, the card and detail dates.

THE PUNCTUATION IS INTL'S, NOT OURS, and that is why this is ONE `format()`
call rather than parts glued with spaces:

  en  Sunday 13 September          (no comma)
  cs  Neděle 13. září              (the dot after the day is REQUIRED)
  ru  Воскресенье, 13 сентября     (comma, genitive month)
  uk  Неділя, 13 вересня           (comma, genitive)

Hand-assembling `${weekday} ${day} ${month}` yields `neděle 13 září` and
`воскресенье 13 сентября` — both wrong, both plausible-looking, and both what a
non-native reader notices first. The ICU data already knows; ask it for the
whole string. Only the leading capital is ours, because the owner asked for
`Neděle` and these are headings rather than prose.

`formatGameDateTime` composes the date half through `formatGameDate` and
appends the clock, rather than asking Intl for both — a combined request drags
in each locale's date-time connector (` v `, ` в `) where the product has always
written a bare `18:30`.

The spec pins all four strings EXACTLY at a fixed Sunday, unlike round 28's
deliberately loose assertions: the punctuation IS the requirement here, and a
test that tolerates it drifting is not testing what was asked for. It also
asserts the failure as an absence, so a switch back to `short` cannot pass
quietly. Two older assertions that pinned `Thu 16 Jul` invert with the round.

ITEM 2 — "Remove banner" beside "Remove photo", and the wallet pair centred.

A NEW FUNCTION RATHER THAN A WIDER OLD ONE, because neither existing one fits.
`remove_profile_photo(p_player_id)` is the admin's and is about the AVATAR;
teaching it a second target would make one call site mean two acts depending on
an argument. `clear_cover_photo()` takes NO argument — it clears the CALLER's
own banner, so an admin calling it would wipe theirs. The new one names its
subject, and the drill asserts precisely that: the admin's own banner survives.

The event catalog is NOT widened. A removed banner is a `profile_photo_removed`
carrying `kind: 'cover'` — the same act on a different column, and a new type
would mean teaching every reader of that stream about a second one.

Migration is SHAPE-ONLY per the rule set last round; the behaviour is drilled in
`supabase/tests/admin_remove_cover.sql` — 9 assertions including the
authorization refusal, the avatar being untouched, idempotence writing no second
event, and an unknown player raising rather than returning null.

The wallet controls are centred between their dividers, and the two image
buttons now share one. Each renders only when there is something to remove, so
a player with a photo and no banner sees one button rather than one enabled and
one dead.

Unit 685/685, SQL 38/38 ALL PASS, tsc clean.

+++ lib/games/days.ts
@@ -308,12 +308,20 @@ function dayHeading(
   locale: Locale,
 ): string {
   const date = new Date(`${key}T12:00:00Z`);
+  /*
+   * FULL NAMES SINCE ROUND 30. `short` rendered `ne 13 9` for a Czech reader —
+   * the owner's own example of the failure — and the abbreviations degrade
+   * worst in exactly the languages that need them most. One `format()` call
+   * rather than glued parts, so each language keeps its own punctuation: the
+   * Czech dot after the day, the Russian and Ukrainian comma. See
+   * `lib/format.ts`, which now does the same thing for the same reason.
+   */
   const full = capitalise(
     new Intl.DateTimeFormat(DATE_LOCALE[locale], {
       timeZone: DISPLAY_TIME_ZONE,
-      weekday: "short",
+      weekday: "long",
       day: "numeric",
-      month: "short",
+      month: "long",
     }).format(date),
     locale,
   );
